% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCHatMatrix.R
\name{createCHatMatrix}
\alias{createCHatMatrix}
\title{createCHatMatrix}
\usage{
createCHatMatrix(productCol, size)
}
\arguments{
\item{productCol}{productCol}

\item{size}{size}
}
\value{
cHatMatrix
}
\description{
createCHatMatrix
}
\examples{

}
