% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDHatMatrix.R
\name{createDHatMatrix}
\alias{createDHatMatrix}
\title{createDHatMatrix}
\usage{
createDHatMatrix(betaArray, size)
}
\arguments{
\item{betaArray}{betaArray}

\item{size}{size}
}
\value{
dHatMatrix
}
\description{
createDHatMatrix
}
\examples{

}
