% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDiffCMatrixVsCHatMatrix.R
\name{createDiffCMatrixVsCHatMatrix}
\alias{createDiffCMatrixVsCHatMatrix}
\title{createDiffCMatrixVsCHatMatrix}
\usage{
createDiffCMatrixVsCHatMatrix(cMatrix, cHatMatrix)
}
\arguments{
\item{cMatrix}{cMatrix}

\item{cHatMatrix}{cHatMatrix}
}
\value{
diffCMatrixVsCHatMatrix
}
\description{
createDiffCMatrixVsCHatMatrix
}
\examples{

}
