% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDiffDMatrixVsDHatMatrix.R
\name{createDiffDMatrixVsDHatMatrix}
\alias{createDiffDMatrixVsDHatMatrix}
\title{createDiffDMatrixVsDHatMatrix}
\usage{
createDiffDMatrixVsDHatMatrix(dMatrix, dHatMatrix)
}
\arguments{
\item{dMatrix}{dMatrix}

\item{dHatMatrix}{dHatMatrix}
}
\value{
diffDMatrixVsDHatMatrix
}
\description{
createDiffDMatrixVsDHatMatrix
}
\examples{

}
