% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDMatrix.R
\name{createDMatrix}
\alias{createDMatrix}
\title{createDMatrix}
\usage{
createDMatrix(size, sigma, psi)
}
\arguments{
\item{size}{size}

\item{sigma}{sigma}

\item{psi}{psi}
}
\value{
dMatrix
}
\description{
createDMatrix
}
\examples{

}
