% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createLRVof2dArray.R
\name{createLRVof2dArray}
\alias{createLRVof2dArray}
\title{createLRVof2dArray}
\usage{
createLRVof2dArray(array2d)
}
\arguments{
\item{array2d}{array2d}
}
\value{
LRVof2dArray
}
\description{
createLRVof2dArray
}
\examples{

}
