% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createLRVof3dArray.R
\name{createLRVof3dArray}
\alias{createLRVof3dArray}
\title{createLRVof3dArray}
\usage{
createLRVof3dArray(array3d)
}
\arguments{
\item{array3d}{array3d}
}
\value{
LRVof3dArray
}
\description{
createLRVof3dArray
}
\examples{

}
