% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createOriginal.R
\name{createOriginal}
\alias{createOriginal}
\title{createOriginal}
\usage{
createOriginal(size = 10, psi = 0.5, sigma = 1)
}
\arguments{
\item{size}{size}

\item{psi}{psi}

\item{sigma}{sigma}
}
\value{

}
\description{
createOriginal
}
\examples{

}
