% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createOriginalMA1.R
\name{createOriginalMA1}
\alias{createOriginalMA1}
\title{createOriginalMA1}
\usage{
createOriginalMA1(size, psi, sigma)
}
\arguments{
\item{size}{size}

\item{psi}{psi}

\item{sigma}{sigma}
}
\value{
originalMA1
}
\description{
createOriginalMA1
}
\examples{

}
