% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createOriginalMA1Replicated.R
\name{createOriginalMA1Replicated}
\alias{createOriginalMA1Replicated}
\title{createOriginalMA1Replicated}
\usage{
createOriginalMA1Replicated(replicationCount = 10, size = 10, psi = 0.5,
  sigma = 1)
}
\arguments{
\item{replicationCount}{replicationCount}

\item{size}{size}

\item{psi}{psi}

\item{sigma}{sigma}
}
\value{
originalMA1Replicated
}
\description{
createOriginalMA1Replicated
}
\examples{

}
