% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createProductCol.R
\name{createProductCol}
\alias{createProductCol}
\title{createProductCol}
\usage{
createProductCol(original)
}
\arguments{
\item{original}{original}
}
\value{
productCol
}
\description{
createProductCol
}
\examples{

}
