% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createProductColReplicated.R
\name{createProductColReplicated}
\alias{createProductColReplicated}
\title{createProductColReplicated}
\usage{
createProductColReplicated(originalReplicated, size, replicationCount)
}
\arguments{
\item{originalReplicated}{originalReplicated}

\item{size}{size}

\item{replicationCount}{replicationCount}
}
\value{
productColReplicated
}
\description{
createProductColReplicated
}
\examples{

}
