% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawCvsCHat.R
\name{drawCvsCHat}
\alias{drawCvsCHat}
\title{drawCvsCHat}
\usage{
drawCvsCHat(cMatrix, cHatMatrix)
}
\arguments{
\item{cMatrix}{cMatrix}

\item{cHatMatrix}{cHatMatrix}
}
\value{
picture
}
\description{

}
\examples{

}
