% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawDvsDHat.R
\name{drawDvsDHat}
\alias{drawDvsDHat}
\title{drawDvsDHat}
\usage{
drawDvsDHat(dMatrix, dHatMatrix)
}
\arguments{
\item{dMatrix}{dMatrix}

\item{dHatMatrix}{dHatMatrix}
}
\value{
picture
}
\description{
drawDvsDHat
}
\examples{

}
