% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawProductCol.R
\name{drawProductCol}
\alias{drawProductCol}
\title{drawProductCol}
\usage{
drawProductCol(productCol, size, mean, sigma, lagCount,
  darkestColor = "dark blue", lightestColor = "white")
}
\arguments{
\item{productCol}{productCol}

\item{size}{size}

\item{mean}{mean}

\item{sigma}{sigma}

\item{lagCount}{lagCount}

\item{darkestColor}{darkestColor}

\item{lightestColor}{lightestColor}
}
\value{
A JPEG file located at \code{"./plots"} directory.
}
\description{
Drawing a plot of \code{ProductCol} values. Exporting it to a JPEG file saved to \code{"./plots"} directory.
}
\examples{
original <- createOriginalMA1(size = 10, psi = 0.5, sigma = 1)
productCol <- createProductCol(original)
drawProductCol(productCol, size = 10, mean = 0, sigma = 1, lagCount = 2, darkestColor = "dark blue", lightestColor = "white"

}
