% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportOriginal.R
\name{exportOriginal}
\alias{exportOriginal}
\title{exportOriginal}
\usage{
exportOriginal(original)
}
\arguments{
\item{original}{original}
}
\value{
an CSV file
}
\description{
exportOriginal
}
\examples{

}
