% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveProductCol.R
\name{saveProductCol}
\alias{saveProductCol}
\title{saveProductCol}
\usage{
saveProductCol(dataframe)
}
\arguments{
\item{dataframe}{dataframe}
}
\value{
a CSV file
}
\description{
saveProductCol
}
\examples{

}
