% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCMatrix.R
\name{createcMatrix}
\alias{createcMatrix}
\alias{createCMatrix}
\title{createcMatrix}
\usage{
createCMatrix(size, sigma, psi)
}
\arguments{
\item{size}{size}

\item{sigma}{sigma}

\item{psi}{psi}
}
\value{
cMatrix
}
\description{
createcMatrix
}
\examples{

}
