% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportOriginal.R
\name{exportOriginal}
\alias{exportOriginal}
\title{exportOriginal}
\usage{
exportOriginal(original)
}
\arguments{
\item{original}{original}
}
\value{
A CSV file located in \code{./export} directory.
}
\description{
Exporting one- or two-dimensional array \code{original} to a CSV file saved to \code{./export} directory.
}
\examples{
original1d <- createOriginalMA1(size = 10, psi = 0.5, sigma = 1)
exportOriginal(original1d)
original2d <- createOriginalMA1Replicated(replicationCount = 10, size = 10, psi = 0.5, sigma = 1)
exportOriginal(original2d)

}
